test: separando testes de relacionamento e estrutura por arquivo
